# docs/swagger.yaml
definitions:
  api.DebtResponse:
    properties:
      amount:
        type: number
      currency:
        type: string
      id:
        type: integer
      owedByUser:
        $ref: '#/definitions/api.UserSummary'
      owedToUser:
        $ref: '#/definitions/api.UserSummary'
    type: object
  api.Exchange:
    properties:
      date:
        type: string
      fromCurrency:
        type: string
      rate:
        type: number
      toCurrency:
        type: string
    type: object
...
  api.SettleDebtsRequest:
    properties:
      owedBy:
        type: string
      owedTo:
        type: string
    required:
    - owedBy
    type: object
  api.SettleDebtsResponse:
    properties:
      owedBy:
        type: string
      owedTo:
        type: string
      settledCount:
        type: integer
      settlements:
        items:
          $ref: '#/definitions/api.SettlementResponse'
        type: array
    type: object
  api.SettlementResponse:
    properties:
      amount:
        type: number
      currency:
        type: string
      date:
        type: string
      id:
        type: integer
      owedByUser:
        $ref: '#/definitions/api.UserSummary'
      owedToUser:
        $ref: '#/definitions/api.UserSummary'
      reversedAt:
        type: string
    type: object
...
  api.UserSummary:
    properties:
      discordHandle:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/v1/admin/exchange-rate:
    get:
      description: Returns the exchange rate for `from` -> `to` on a given date (defaults
        to today if date is omitted).
      parameters:
      - description: Base currency code (e.g. USD)
        in: query
        name: from
        required: true
        type: string
      - description: Target currency code (e.g. EUR)
        in: query
        name: to
        required: true
        type: string
      - description: Date in YYYY-MM-DD format
        in: query
        name: date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Exchange'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get exchange rate between two currencies
      tags:
      - admin
  /api/v1/admin/exchange-rates:
    get:
      description: Returns the exchange rates stored in the DB
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.Exchange'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all exchange rates stored in the DB
      tags:
      - admin
  /api/v1/debts:
    get:
      parameters:
      - default: USD
        description: Currency for generated debts
        in: query
        name: currency
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.DebtResponse'
            type: array
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Fetch all debts as From-To user debt pairs
      tags:
      - debts
  /api/v1/debts/settle:
    post:
      consumes:
      - application/json
      parameters:
      - description: Settle debts payload
        in: body
        name: settle
        required: true
        schema:
          $ref: '#/definitions/api.SettleDebtsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SettleDebtsResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Settle debts for a user
      tags:
      - debts
  /api/v1/debts/settlements:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.SettlementResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Fetch settlement history
      tags:
      - debts
  /api/v1/debts/settlements/{id}/reverse:
    post:
      parameters:
      - description: Settlement ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SettlementResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Reverse a settlement by ID
      tags:
      - debts
  /api/v1/debts/users:
    get:
      parameters:
      - default: USD
        description: Currency for generated debts
        in: query
        name: currency
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/api.DebtResponse'
              type: array
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Fetch a map of users to their debts
      tags:
      - debts
...
swagger: "2.0"